pyproject.toml
[project]
name = "nbsapi"
authors = [
    { name="Stephan Hügel", email="urschrei@gmail.com" }
]
requires-python = ">=3.12"
license = {text = "MIT License"}
keywords = ["Geo"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: GIS"
    ]
version = "0.2.0"
description = "An API for nature-based solutions"
readme = "README.md"
dependencies = [
    "alembic>=1.13.2",
    "bcrypt>=4.2.1",
    "click>=8.1.7",
    "email-validator>=2.2.0",
    "fastapi>=0.115.8",
    "geoalchemy2>=0.15.2",
    "geojson>=3.1.0",
    "greenlet>=3.1.0",
    "httpx>=0.27.2",
    "passlib>=1.7.4",
    "psycopg>=3.2.1",
    "pydantic-settings>=2.5.2",
    "pyjwt>=2.9.0",
    "python-multipart==0.0.18",
    "shapely>=2.0.6",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.0",
]
[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "tavern>=2.15.0",
]

[project.scripts]
nbsapi = "nbsapi.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--tavern-global-cfg=tests/common.yaml",
    "--tb=short"
]
markers = [
    "auth: marks tests that require authentication",
    "public: marks tests that don't require authentication"
]
filterwarnings = [
    "ignore:cannot collect test class 'TestType' because it has a __init__ constructor",
    "ignore:Can't validate argument 'spatial_index'; can't locate any SQLAlchemy dialect named 'spatial'",
    "ignore:ast.Str is deprecated and will be removed in Python 3.14; use ast.Constant instead:DeprecationWarning",
    "ignore:ast.Num is deprecated and will be removed in Python 3.14; use ast.Constant instead:DeprecationWarning",
    "ignore:ast.NameConstant is deprecated and will be removed in Python 3.14; use ast.Constant instead:DeprecationWarning",
    "ignore:Attribute s is deprecated and will be removed in Python 3.14; use value instead:DeprecationWarning"
]

[tool.hatch.build.targets.wheel]
packages = ["src/nbsapi"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["F", "E", "W", "N", "D4", "UP", "PL", "TRY", "PERF", "FURB", "RUF", "S", "DTZ"]
ignore = ["ANN001", "ANN003", "ANN202", "ANN201", "DOC201", "E501", "PLR0904", "PLR0913", "PLR0917", "SLF001", "FIX002", "D400", "D415", "PLR2004", "S101", "N815", "TRY003", "PLR0912", "S311", "PLR0915"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
